refactor: error 모달 통합 및 alias 적용
