+++ app/(anon)/interview/components/PendingInterviewsList.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation';
 import axios from 'axios';
 import { NoneReports } from './NoneReports';
 import { LoadingSpinner } from '@/app/(anon)/components/LoadingSpinner';
-import LoginModal from '../../components/LoginModal';
+import LoginModal from '@/app/(anon)/components/modal/LoginModal';
 import { useModalStore } from '@/stores/useModalStore';
 import Arrow from '@/public/assets/icons/arrow-right.svg';
