+++ app/(anon)/interview/components/modal/GenerateQuestionModal.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { useRouter } from 'next/navigation';
-import Modal from '../../../components/Modal';
-import ModalOverlay from '../../../components/ModalOverlay';
+import Modal from '@/app/(anon)/components/modal/Modal';
+import ModalOverlay from '@/app/(anon)/components/modal/ModalOverlay';
 import Check from '@/public/assets/icons/check-circle.svg';
 import { useModalStore } from '@/stores/useModalStore';
 import { useReportStore } from '@/stores/useReportStore';
